Updated RVsharing vignette.
The method RVsharing works for both classes for siblings and first cousins.  Neither class works for second cousin.

It is not clear to me what the missing data code is for RVsharing.  It seems that NA is working, but the code makes it look like it might be a zero.