In trying to get through R check I see only two troublesome messages.  See below.
~~

Warning message:
.onUnload failed in unloadNamespace() for ‘Bureau’, details:
  call: library.dynam.unload(THISPKG, libpath)
  error: shared object ‘Bureau.so’ was not loaded

~~

PFU.fromphi: no visible binding for global variable ‘Nf’

~~

Not sure what to do about the first yet, but the second is easy to fix.  I've emailed Alexandre to bring it to his attention.

~~

All the necessary Rd files now exist, but they have not been "filled in" yet.  Think of them right now as templates to help us get through the check, and then we can go back and fill them in as needed.