Updated RVsharing vignette.
Now in the vignette we simulate 10,000 realizations of the gene-drop with at least one RV observed in the descendants.

Note also that the name of the vignettes directory was changed from vignette to vignettes.  I believe this is what is expected, and that the original name vignette was a mistake.