Changed signature for RVsharing method.
Now the user may input a pedigree object or three vectors.  The awkward part is that the first argument must share a name if it is to work without naming the first argument.  The best I can do for a shared name between id and ped is data.  I will continue to look for a better solution..