Updated ped2trio function.
This function transforms a kinship2 pedigree object to a recursive trio object for use in the recursive gene-dropping algorithm.  Alexandre wrote the section that transforms a pedigree object to a recursive list of trios.  Sam wrote the bit that converts the trio list object to a RVsharing trio object.

Right now it works for first cousins, and is likely to not work for other pedigrees.  I will examine example pedigrees one by one and deal with issues as they arise.