Fixed error in ped2trio.R.
Now it works for both first and second cousins.  It does now also reuqire that the user enter ped.vec as an argument to ped2trio.  ped.vec is avector that contains the number of trios in each level of the pedigree.  For our first cousin example it is c(2,1) and for our second cousin example it is c(2,2,1).

There is a strange reversal of trio IDs in the intermediate level of the second cousin pedigree.  Hopefully, it will not cause trouble.

I am certain that this function is fully general yet, and it will only work for pedigrees with up to three levels.  I will look at some more complicated pedigrees and see how it holds up.