Corrected the problem in GeneDrop.
Somehow in the creation of the trio object the structure of the class was changed unexpectedly.  The offspring slot initially was a list of trio objects.  Now it is a list of one-element lists that contain the trio object.  So, we needed an additional [[1]] to retreive the trio object.  This is a strange structure, and is no doubt unintentional.  Hopefully we can clean it up.  But, for now, we can proceed like this.