Updated GeneDropSim.fn and wrote documentation
The first two arguments of GeneDropSim.fn are now a list of trio objects
and a vector of IDs of the pedigree members. The code has been changed
to loops over the list of trio objects. I documented the arguments and
value of the function and gave an example, but details remain to be
written.