Commented out line in ComputeKinshipPropCoef example
There was a problem with the example in ComputeKinshipPropCoef so I
commented out the offending line until Alexandre has a look.  The
problem was that there was discordance between the object expected in
the function and the object passed to the function in the example.
More specifically, ComputeKinshipPropCoef wants the list returned by
RVsharing, while the example passed a kinship2 pedigree object.