Commented out function with library.dynam.unload from zzz.R
There seems to be a bug in library.dynam.unload that causes a warning
on check.  To be on the safe side I simply commented out the offending
code from zzz.R.  I don't understand what library.dynam.unload does
anyway, and so it is best to not use it.